Revert "refactor: centralizar Navbar con soporte de roles y barra de búsqueda"
