+++ biblioteca-alejandria/app/page.tsx
@@ -1,9 +1,6 @@
-import Navbar from "@/components/Navbar";
-
 export default function Home() {
   return (
-    <div className="min-h-screen bg-brand-bg flex flex-col">
-      <Navbar />
+    <div>
     </div>
   );
 }
